Fix anchor links in MDX files when prettyUrls is false.
If the app is configured with `prettyUrls: false`, then anchor links in
MDX files may not behave as expected. For example, a link like:

[foo](#bar)

Will be generated in the output as:

<a href="#bar.html">foo</a>

Instead of remaining `href="#bar"`, like expected (since the anchor of
'#bar.html` probably won't exist on the page).

Similarly, a link to `/foo#bar` will get translated to `/foo#bar.html`
instead of what I think would be intended by how other URLs get
rewritten (in which case, I think `/foo.html#bar` would be expected).

This fixes the issue by skipping the MDX URL rewriting when anchor tags
on the current page (no path is present) are encountered. For anchor
links on other pages, the anchor is split from the path, and then the
existing path rules are used, adding the anchor back in afterwards.

Using a URL parsing library, instead of simply splitting the string by
`#` may be more robust, but Node.js's `URL` class doesn't handle
relative links or fragment-only links, so it won't work for this
use-case. But I think this simple splitting approach will hopefully meet
the needs here.