fix: avoid adding tsconfig.json to the published package (close #87)
Internally, vite will use certain properties in tsconfig.json to ensure
the code is transformed with esbuild as expected.

The problem is that the `dist/client/tsconfig.json` file included in the
package was referencing `tsconfig.json` using relative paths, but that
file is not included.

Since iles does not use any of these special properties, it's better not
to ship client/tsconfig.json, which is only meant for internal use.