fix: prevent hmr errors when using islands with multiple root nodes (fragments) (close #108)
Before, when rendering Vue islands two instances could live in the same
element, the first when prerendering the island, the second when the
hydration script mounted on the element.

By prerendering the island using a static vnode, no live instances are
created, preventing HMR from running twice on the same element which can
trigger an error in `removeFragment`.

A simpler alternative would be to avoid prerendering during dev, but the
visual flicker that happens between page navigations is annoying.