fix: prevent duplicated slugs for headings with the same title
A counter is used to track the number of duplicated titles.

To pass the counter to the slug generator, the API of generateSlug is
modified.

ids are taken into account when generating distinct slugs, but
duplicated ids are not resolved and kept as-is, because it's the user's
responsibility to not set duplicated ids.

An example (test suite):

    title     id      slug
    foo       none    foo
    bar       none    bar-2
    foo       none    foo-2
    foo       bar     bar
    foo-2     none    foo-2-2
    foo-4     none    foo-4
    foo       none    foo-3
    foo       none    foo-5
    bar       none    bar-3
    foo-2     none    foo-2-3
    foo-2-2   none    foo-2-2-2
    foo       bar     bar